Traffic Light Formatter

Traffic Light Formatter

Usage

set_formatter_traffic_light(
  widget,
  column,
  min = NA,
  max = NA,
  color = c("green", "orange", "red"),
  hoz_align = "center"
)

Arguments

widget

(tabulator) A tabulator widget.

column

(character): The column the formatter is applied to.

min

(numeric): The minimum value for progress bar. If set to NA the minimum value of the column is used.

max

description

color

(character): Either a single color or a vector of colors

hoz_align

(character): The horizontal alignment of the column.