Create a Tabulator Widget

Dots in column names are replaced by underscores.

Usage

tabulator(
  data,
  options = tabulator_options(),
  editable = FALSE,
  sheetjs = FALSE,
  theme = c("default", "midnight", "modern", "simple", "site", "bootstrap3",
    "bootstrap4", "bootstrap5", "bulma", "materialize", "semanticui"),
  width = NULL,
  height = NULL,
  element_id = NULL,
  ...
)

Arguments

data

(data.frame, character or list): In spreadsheet mode data needs to be a list or NULL for an empty spreadsheet.

options

(list): Setup options. See tabulator_options.

editable

(bool): Whether the table is editable.

sheetjs

(bool): Whether to add sheetjs (https://sheetjs.com/) dependency, which is needed for xlsx downloads.

theme

(character): Theme to apply to the table.

width

Width of the widget.

height

Height of the widget.

element_id

The unique ID of the widget.

...

Named arguments that are appended to the options parameter.

Examples

setup <- tabulator_options(
  spreadsheet = TRUE,
  spreadsheet_sheets = list(
    list(
      title = "First",
      data = list(c(1, 2, 3))
    ),
    list(
      title = "Second",
      data = list(c(4, 5, 6))
    )
  ),
  spreadsheet_sheet_tabs = TRUE
)

tabulator(data = NULL, setup, theme = "midnight")