Tabulator Setup Options

Tabulator Setup Options

Usage

tabulator_options(
  height = "311px",
  history = FALSE,
  columns = NULL,
  layout = c("fitColumns", "fitData", "fitDataFill", "fitDataStretch", "fitDataTable"),
  responsive_layout = FALSE,
  movable_columns = FALSE,
  header_visible = TRUE,
  row_height = NULL,
  add_row_pos = c("bottom", "top"),
  movable_rows = FALSE,
  resizable_rows = FALSE,
  frozen_rows = NULL,
  index = "id",
  group_by = NULL,
  edit_trigger_event = "dblclick",
  selectable_rows = "highlight",
  pagination = FALSE,
  pagination_size = 10,
  pagination_add_row = "page",
  spreadsheet = FALSE,
  spreadsheet_rows = NULL,
  spreadsheet_columns = NULL,
  spreadsheet_column_definition = NULL,
  spreadsheet_sheets = NULL,
  spreadsheet_sheet_tabs = NULL,
  ...
)

Arguments

height

(character) The height of the table in px.

history

(bool): Enable user interaction history functionality.

columns

(list): Column definitions. If set to NULL definitions are auto generated from table data.

layout

(character): Layout mode of the table columns.

responsive_layout

(bool): Automatically hide and show columns to fit the width of the Tabulator element.

movable_columns

(bool): Allow users to move and reorder rows.

header_visible

(bool): Whether header is visible.

row_height

(numeric): A fixed height of the rows.

add_row_pos

(character): The table position where new rows are added.

movable_rows

(bool): Allow users to move and reorder rows.

resizable_rows

(bool): Allow user to resize rows.

frozen_rows

(numeric): Number of frozen rows.

index

(character): Field to be used as the unique index for each row.

group_by

(character): Field to group rows by.

edit_trigger_event

description

selectable_rows

description

pagination

(bool): description

pagination_size

(integer): description

pagination_add_row

(character): description

spreadsheet

(bool): description

spreadsheet_rows

(integer): description

spreadsheet_columns

(integer): description

spreadsheet_column_definition

description

spreadsheet_sheets

(list): description

spreadsheet_sheet_tabs

(bool): description

...

Further options.

See also