Create a tabulator widget

Create a tabulator widget

Usage

tabulator(
  data,
  options = tabulator_options(),
  editable = FALSE,
  luxon = FALSE,
  sheetjs = FALSE,
  theme = c("default", "midnight", "modern", "simple", "site", "bootstrap3",
    "bootstrap4", "bootstrap5", "bulma", "materialize", "semanticui"),
  col_select = NULL,
  width = NULL,
  height = NULL,
  element_id = NULL,
  ...
)

Arguments

data

(data.frame, character or list): In spreadsheet mode data needs to be a list or NULL for an empty spreadsheet.

options

(list): Setup options. See tabulator_options.

editable

(bool): Whether the table is editable.

luxon

(bool): Whether to add luxon dependency, which is required for set_formatter_datetime.

sheetjs

(bool): Whether to add sheetjs dependency, which is required for trigger_download to support Excel type downloads.

theme

(character): Theme to apply to the table.

col_select

(character vector) Columns to select. Passed to read_csv if data is a file name.

width

Width of the widget.

height

Height of the widget.

element_id

The unique ID of the widget.

...

Named arguments that are appended to the options parameter.

Value

tabulator htmlwidget

Details

Dots in column names are replaced by underscores.

Examples

setup <- tabulator_options(
  spreadsheet = TRUE,
  spreadsheet_sheets = list(
    list(
      title = "First",
      data = list(c(1, 2, 3))
    ),
    list(
      title = "Second",
      data = list(c(4, 5, 6))
    )
  ),
  spreadsheet_sheet_tabs = TRUE
)

tabulator(data = NULL, setup, theme = "midnight")