Tabulator Setup Options

Tabulator Setup Options

Usage

tabulator_options(
  height = "311px",
  history = FALSE,
  columns = NULL,
  layout = c("fitColumns", "fitData", "fitDataFill", "fitDataStretch", "fitDataTable"),
  responsive_layout = FALSE,
  movable_columns = FALSE,
  header_visible = TRUE,
  row_height = NULL,
  add_row_pos = c("bottom", "top"),
  movable_rows = FALSE,
  resizable_rows = FALSE,
  frozen_rows = NULL,
  index = "id",
  group_by = NULL,
  group_start_open = TRUE,
  group_toggle_element = NULL,
  edit_trigger_event = c("dblclick", "click", "focus"),
  selectable_rows = "highlight",
  pagination = FALSE,
  pagination_size = 10,
  pagination_size_selector = FALSE,
  pagination_add_row = c("page", "table"),
  spreadsheet = FALSE,
  spreadsheet_rows = NULL,
  spreadsheet_columns = NULL,
  spreadsheet_column_definition = NULL,
  spreadsheet_sheets = NULL,
  spreadsheet_sheet_tabs = NULL,
  ...
)

Arguments

height

(character) The height of the table in px.

history

(bool): Enable user interaction history functionality.

columns

(list): Column definitions. If set to NULL definitions are auto generated from table data.

layout

(character): Layout mode of the table columns.

responsive_layout

(bool): Automatically hide and show columns to fit the width of the Tabulator element.

movable_columns

(bool): Allow users to move and reorder rows.

header_visible

(bool): Whether header is visible.

row_height

(numeric): A fixed height of the rows.

add_row_pos

(character): The table position where new rows are added.

movable_rows

(bool): Allow users to move and reorder rows.

resizable_rows

(bool): Allow user to resize rows.

frozen_rows

(numeric): Number of frozen rows.

index

(character): Field to be used as the unique index for each row.

group_by

(character vector): Field to group rows by. Pass multiple values for multi level grouping.

group_start_open

(bool vector): Open state of groups when grouping is enabled. The length of the vector depends on the number of levels set with group_by.

group_toggle_element

(character, bool): One of arrow or header. Set to FALSE to disable toggling at all.

edit_trigger_event

(character): Event that triggers a cell edit.

selectable_rows

(character, bool, integer) Set to FALSE to disble row selection. If set to TRUE you can select as many rows as you want. If set to "highlight" rows are just highlighted but do not change state when clicked. An integer value sets the maximum number of rows that can be selected.

pagination

(bool): Whether to enable pagination.

pagination_size

(integer): Number of rows on each page.

pagination_size_selector

(list): Add pagination size selector.

pagination_add_row

(character): Where to add rows to the table when pagination is enabled.

spreadsheet

(bool): Whether to enable spreadsheet mode.

spreadsheet_rows

(integer): Number of spreadsheet rows.

spreadsheet_columns

(integer): Number of spreadsheet columns.

spreadsheet_column_definition

Column definition used for all columns in the sheet.

spreadsheet_sheets

(list): List of sheet definitions.

spreadsheet_sheet_tabs

(bool): Whether to show sheet tabs in the footer.

...

Further options.

See also

Examples

setup <- tabulator_options(
  group_by = "Species",
  pagination = TRUE,
  pagination_size = 10,
  pagination_size_selector = c(10, 20, 50),
  movable_columns = TRUE,
  movable_rows = TRUE
)
tabulator(iris, setup)